+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/mining/miningcharge/MiningChargeRenderer.java
@@ -1,7 +1,6 @@
 package insane96mcp.iguanatweaksexpanded.module.mining.miningcharge;
 
 import com.mojang.blaze3d.vertex.PoseStack;
-import com.mojang.math.Axis;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.block.BlockRenderDispatcher;
 import net.minecraft.client.renderer.entity.EntityRenderer;
@@ -10,6 +9,7 @@ import net.minecraft.client.renderer.entity.TntMinecartRenderer;
 import net.minecraft.client.renderer.texture.TextureAtlas;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.Mth;
+import net.minecraft.world.level.block.state.BlockState;
 
 public class MiningChargeRenderer extends EntityRenderer<PrimedMiningCharge> {
     private final BlockRenderDispatcher blockRenderer;
@@ -32,10 +32,9 @@ public class MiningChargeRenderer extends EntityRenderer<PrimedMiningCharge> {
             pPoseStack.scale(f1, f1, f1);
         }
 
-        pPoseStack.mulPose(Axis.YP.rotationDegrees(-90.0F));
-        pPoseStack.translate(-0.5F, -0.5F, 0.5F);
-        pPoseStack.mulPose(Axis.YP.rotationDegrees(90.0F));
-        TntMinecartRenderer.renderWhiteSolidBlock(this.blockRenderer, MiningCharge.MINING_CHARGE.block().get().defaultBlockState(), pPoseStack, pBuffer, pPackedLight, fuse / 5 % 2 == 0);
+        BlockState state = MiningCharge.MINING_CHARGE.block().get().defaultBlockState().setValue(MiningChargeBlock.FACING, pEntity.getDirection());
+        pPoseStack.translate(-0.5F, -0.5F, -0.5F);
+        TntMinecartRenderer.renderWhiteSolidBlock(this.blockRenderer, state, pPoseStack, pBuffer, pPackedLight, fuse / 5 % 2 == 0);
         pPoseStack.popPose();
         super.render(pEntity, pEntityYaw, pPartialTick, pPoseStack, pBuffer, pPackedLight);
     }